Integra API raspay

+++ src/main/java/com/client/ws/rasmooplus/dto/wsraspay/OrderDto.java
@@ -2,11 +2,13 @@ package com.client.ws.rasmooplus.dto.wsraspay;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record OrderDto(
     String id,
     String customerId,
     String productAcronym,
-    Long discount
+    BigDecimal discount
 ) {
 }
